attempt to render friends list only on user show
Previously would attempt to check for motimate list on every loaded page
leading to console errors. While this did not affect the user experience
(at least at the moment), it is an unneccessary check that is an easy
fix.

Will not only attempt to populate list if it's container is found on the
page.